Share setup code between datastore and datastoreFits tests
There was a copy and paste error in datastoreFits and this
code sharing fixes that so we do not reuse the same
directory.
